move snap l10n extensions into separate file
