+++ lib/snapd.dart
@@ -1,4 +1,5 @@
 export 'src/snapd/snap_category_enum.dart';
+export 'src/snapd/snap_l10n.dart';
 export 'src/snapd/snap_launcher.dart';
 export 'src/snapd/snap_model.dart';
 export 'src/snapd/snapd_service.dart';
